register modal with three registration views
