+++ src/components/nav/NavBar.js
@@ -1,10 +1,13 @@
-import { useRef } from "react";
+import { useEffect, useRef, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
+import { ArtistRegister } from "../auth/ArtistRegister";
+import { Register } from "../auth/Register";
 import "./NavBar.css";
-
 // Creates and exports NavBar component
 // uses Props in the argument defined in Rare.js
 export const NavBar = ({ token, setToken }) => {
+    const [registerState, setRegisterState] = useState("register")
+
     //defines navigate variable to use useNavigate hook
     const navigate = useNavigate();
     /*
@@ -12,6 +15,58 @@ export const NavBar = ({ token, setToken }) => {
     •Defines variable where useRef is set to an initial value of null
     */
     const navbar = useRef();
+    useEffect(() => {
+        document.addEventListener('click', () => {
+            
+            // Functions to open and close a modal
+            function openModal($el) {
+                $el.classList.add('is-active')
+                setRegisterState("register")
+                ;
+            }
+
+            function closeModal($el) {
+                $el.classList.remove('is-active')
+                setRegisterState("register")
+                ;
+            }
+
+            function closeAllModals() {
+                (document.querySelectorAll('.modal') || []).forEach(($modal) => {
+                    closeModal($modal);
+                });
+            }
+
+            // Add a click event on buttons to open a specific modal
+            (document.querySelectorAll('.js-modal-trigger') || []).forEach(($trigger) => {
+                const modal = $trigger.dataset.target;
+                const $target = document.getElementById(modal);
+
+                $trigger.addEventListener('click', () => {
+                    openModal($target)
+                });
+            });
+
+            // Add a click event on various child elements to close the parent modal
+            (document.querySelectorAll('.modal-background, .modal-close, .modal-cancel, .modal-card-head .delete, .modal-card-foot .button .close') || []).forEach(($close) => {
+                const $target = $close.closest('.modal');
+
+                $close.addEventListener('click', () => {
+                    closeModal($target)
+                    
+                });
+            });
+
+            // Add a keyboard event to close all modals
+            document.addEventListener('keydown', (event) => {
+                const e = event || window.event;
+
+                if (e.keyCode === 27) { // Escape key
+                    closeAllModals();
+                }
+            });
+        });
+    }, [])
     /*
     the mobile icon used for opening a menu
     hidden until user adjusts width
@@ -24,7 +79,8 @@ export const NavBar = ({ token, setToken }) => {
         hamburger.current.classList.toggle("is-active");
         navbar.current.classList.toggle("is-active");
     };
-    return (
+    return (<>
+
         <nav
             className="navbar"
             role="navigation"
@@ -93,9 +149,9 @@ export const NavBar = ({ token, setToken }) => {
                                     </button>
                                 ) : (
                                     <>
-                                        <Link to="/register" className="button is-rounded is-info">
+                                        <button className="js-modal-trigger button is-rounded is-info" data-target="modal-js-example">
                                             Register
-                                        </Link>
+                                        </button>
                                         <Link to="/login" className="button is-rounded">
                                             Login
                                         </Link>
@@ -106,7 +162,16 @@ export const NavBar = ({ token, setToken }) => {
                     </div>
                 </div>
             </div>
+            <div id="modal-js-example" class="modal">
+                <div class="modal-background"></div>
+
+                        <Register setRegisterState={setRegisterState} registerState={registerState}/>
+                
+
+                <button class="modal-close is-large" aria-label="close"></button>
+            </div>
         </nav>
+    </>
     );
 };
