+++ src/index.css
@@ -1,15 +1,16 @@
+@charset "utf-8";
+
+@import url('https://fonts.googleapis.com/css?family=Playfair+Display');
+
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  font-family: "Playfair Display", -apple-system, BlinkMacSystemFont, serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+  font-family: "Playfair Display", source-code-pro, serif;
 }
 
 /* imports Bulma framework for css */
@@ -18,6 +19,13 @@ a {
   color: #309875;
 }
 a:hover { text-decoration: none; color:lightgray; }
+
+body {
+  margin: 0;
+  font-family: "Playfair Display", -apple-system, BlinkMacSystemFont, serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
 
 html, body {
     background-color: #131313;
@@ -50,4 +58,8 @@ html, body {
 a.navbar-item:focus, a.navbar-item:focus-within, a.navbar-item:hover, a.navbar-item.is-active, .navbar-link:focus, .navbar-link:focus-within, .navbar-link:hover, .navbar-link.is-active {
   background-color: #FB9970;
   color: #6295AA;
+}
+
+input, textarea {
+  font-family: "Playfair Display";
 }
